Restructured Application
Restructured application to make it more modular.
