+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
-
-import { App } from './App';
-import './styles.css';
+import { App } from './modules/App/index.js';
+import './styles/styles.css';
 
 const rootDiv = document.createElement('div');
 rootDiv.id = 'root';
