+++ src/modules/App/index.js
@@ -1,0 +1,1 @@
+export { App } from './component';
