+++ src/modules/Todo/index.js
@@ -1,0 +1,1 @@
+export { Todo } from './component';
